Проверочка

+++ Database_scripts/ViewTest.sql
@@ -1,4 +1,6 @@
-DROP VIEW GameRound_User
+GO
+IF OBJECT_ID('GameRound_User', 'V') IS NOT NULL
+	DROP VIEW GameRound_User
 GO
 CREATE VIEW GameRound_User AS
 	SELECT GameRound.tournament_id, GameRound.gamenumber, GameRound.id
